1.2.0 fixes

+++ app/src/main/java/org/dhis2/data/forms/dataentry/AttributeValueStore.java
@@ -3,8 +3,6 @@ package org.dhis2.data.forms.dataentry;
 import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteStatement;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 
 import com.squareup.sqlbrite2.BriteDatabase;
 
@@ -23,6 +21,8 @@ import java.util.Objects;
 
 import javax.annotation.Nonnull;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import io.reactivex.BackpressureStrategy;
 import io.reactivex.Flowable;
 
@@ -139,9 +139,11 @@ public final class AttributeValueStore implements DataEntryStore {
             }
             String eventUid = eventUid(attribute);
             // ToDo: write test cases for different events
-            return (long) briteDatabase.update(TrackedEntityDataValueModel.TABLE, dataValue,
-                    TrackedEntityDataValueModel.Columns.DATA_ELEMENT + " = ? AND " +
-                            TrackedEntityDataValueModel.Columns.EVENT + " = ?", attribute == null ? "" : attribute, eventUid == null ? "" : eventUid);
+            if (!isEmpty(eventUid))
+                return (long) briteDatabase.update(TrackedEntityDataValueModel.TABLE, dataValue,
+                        TrackedEntityDataValueModel.Columns.DATA_ELEMENT + " = ? AND " +
+                                TrackedEntityDataValueModel.Columns.EVENT + " = ?", attribute == null ? "" : attribute, eventUid == null ? "" : eventUid);
+            else return -1;
         }
     }
 
@@ -181,8 +183,9 @@ public final class AttributeValueStore implements DataEntryStore {
 
     private long insert(@NonNull String attribute, @NonNull String value, valueType valueType) {
         if (valueType == ATTR) {
-            String created = BaseIdentifiableObject.DATE_FORMAT
-                    .format(Calendar.getInstance().getTime());
+            Date date = Calendar.getInstance().getTime();
+            String created = BaseIdentifiableObject.DATE_FORMAT.format(date);
+/*
 
             sqLiteBind(insertStatement, 1, created == null ? "" : created);
             sqLiteBind(insertStatement, 2, created == null ? "" : created);
@@ -192,13 +195,36 @@ public final class AttributeValueStore implements DataEntryStore {
 
             long inserted = briteDatabase.executeInsert(
                     TrackedEntityAttributeValueModel.TABLE, insertStatement);
-            insertStatement.clearBindings();
+            insertStatement.clearBindings();*/
 
-            return inserted;
+            Cursor teiCursor = briteDatabase.query(SELECT_TEI, enrollment);
+            String teiUid = null;
+
+            try {
+                if (teiCursor.moveToFirst())
+                    teiUid = TrackedEntityInstanceModel.create(teiCursor).uid();
+            } finally {
+                if (teiCursor != null)
+                    teiCursor.close();
+            }
+
+            if (teiUid != null) {
+                TrackedEntityAttributeValueModel attributeValueModel =
+                        TrackedEntityAttributeValueModel.builder()
+                                .created(date)
+                                .lastUpdated(date)
+                                .trackedEntityAttribute(attribute)
+                                .trackedEntityInstance(teiUid)
+                                .value(value)
+                                .build();
+
+                return briteDatabase.insert(TrackedEntityAttributeValueModel.TABLE, attributeValueModel.toContentValues());
+            } else
+                return -1;
         } else {
             Date created = Calendar.getInstance().getTime();
             String eventUid = eventUid(attribute);
-            if(!isEmpty(eventUid)) {
+            if (!isEmpty(eventUid)) {
                 TrackedEntityDataValueModel dataValueModel =
                         TrackedEntityDataValueModel.builder()
                                 .created(created)
@@ -209,7 +235,7 @@ public final class AttributeValueStore implements DataEntryStore {
                                 .build();
                 return briteDatabase.insert(TrackedEntityDataValueModel.TABLE,
                         dataValueModel.toContentValues());
-            }else
+            } else
                 return -1;
         }
     }
@@ -225,10 +251,13 @@ public final class AttributeValueStore implements DataEntryStore {
             return deleted;
         } else {
             String eventUid = eventUid(attribute);
-            return (long) briteDatabase.delete(TrackedEntityDataValueModel.TABLE,
-                    TrackedEntityDataValueModel.Columns.DATA_ELEMENT + " = ? AND " +
-                            TrackedEntityDataValueModel.Columns.EVENT + " = ?",
-                    attribute == null ? "" : attribute, eventUid == null ? "" : eventUid);
+            if (!isEmpty(eventUid))
+                return (long) briteDatabase.delete(TrackedEntityDataValueModel.TABLE,
+                        TrackedEntityDataValueModel.Columns.DATA_ELEMENT + " = ? AND " +
+                                TrackedEntityDataValueModel.Columns.EVENT + " = ?",
+                        attribute == null ? "" : attribute, eventUid == null ? "" : eventUid);
+            else
+                return -1;
         }
 
     }
@@ -247,7 +276,7 @@ public final class AttributeValueStore implements DataEntryStore {
     }
 
     private boolean checkUnique(String attribute, String value) {
-        if(attribute!=null && value!=null) {
+        if (attribute != null && value != null) {
             Cursor uniqueCursor = briteDatabase.query("SELECT TrackedEntityAttributeValue.value FROM TrackedEntityAttributeValue" +
                     " JOIN TrackedEntityAttribute ON TrackedEntityAttribute.uid = TrackedEntityAttributeValue.trackedEntityAttribute" +
                     " WHERE TrackedEntityAttribute.uid = ? AND" +
@@ -261,7 +290,7 @@ public final class AttributeValueStore implements DataEntryStore {
                 uniqueCursor.close();
                 return !hasValue;
             }
-        }else
+        } else
             return true;
     }
